Add settings page with calendar id form and recent events on home page
Also removed the calendar api script. home.js and settings.js are both in need of refactoring to encapsulate some of the logic away to make this more maintanable. So the calendar api class and some sort of form class should be added back in the future.

Also encapsulated some of the logic in background.js (main process) into a startup.js helper file.