chore(git): use conventional changelog form for git messages w/ commitizen
This includes a helper with commitzen to be able to have
prompts for when committing to the repo. While the git
commit style will use commitizen, the changelog will not.
I'm not a huge fan of autogenerated changelogs from git
commits, but I do think having categories and scopes for
commit messages forces you to not only have smaller and
more focused commits but also, I find it more readable
when going through the git history to be able to see at
a glance what types of commits were made.